added few upgrades

+++ qc/main.py
@@ -78,13 +78,18 @@ def main(cpac_output_dir, qc_dir, overlay_csv=False, n_procs=10):
         result_df['file_name'] = result_df.apply(lambda row: gen_filename(res1_row=row), axis=1)
         result_df['plots_dir'] = plots_dir
         result_df['plot_path'] = result_df.apply(lambda row: generate_plot_path(create_directory(row['sub'], row['ses'], row['plots_dir']), row['file_name']), axis=1)
-
+        
+        columns_to_keep = ['sub', 'ses', 'file_path_1', 'file_path_2', 'file_name', 'plots_dir', 'plot_path']
+        result_df = result_df[columns_to_keep].copy()
 
     result_df['relative_path'] = result_df.apply(lambda row: os.path.relpath(row['plot_path'], qc_dir), axis=1)
     
     # save the result_df to csv
     result_df_csv_path = os.path.join(qc_dir, "results.csv")
-    result_df.to_csv(result_df_csv_path, index=False)
+    if os.path.exists(result_df_csv_path):
+        result_df.to_csv(result_df_csv_path, mode='a', header=False, index=False)
+    else:
+        result_df.to_csv(result_df_csv_path, index=False)
     
     args = [
         (
